Fix connection + producer pooling
